Reimplement crate to use inherent impls over traits
* Redesign API as inherent struct implementations with `Deref` and
  `DerefMut` for handling backwards compatibility

* Rename `Version` to `VersionCode`, `ApiVersion` trait to `Version`

* Use a single `Entry` type, since the aliases point to the same struct

* Add `HasPrevious` trait to recursively determine version compatibility
  at compile-time

* Remove the `prelude` module

* Remove the `api` module

* Move `CaptureOptions`, `InputButton`, `OverlayBits` to `settings.rs`

* Move `DeviceHandle` and `WindowHandle` to `handles.rs`

* Remove `RenderDocV###` trait boilerplate code from `renderdoc-derive`
