--- src/prelude.rs
@@ -1,4 +1,0 @@
-//! Contains common types that can be glob-imported (`*`) for convenience.
-
-pub use api::*;
-pub use entry::ApiVersion;
